refactor: re-design new member flow
